S2D-55 refacotr : ClosetService 관심사 분리
 - ClosetService : 옷장에 관련된 로직
 - ClosetItemService : 옷장 안에 아이템에 관련된 로직
