+++ src/main/java/com/sluv/server/domain/closet/service/ClosetService.java
@@ -1,7 +1,5 @@
 package com.sluv.server.domain.closet.service;
 
-import com.sluv.server.domain.closet.dto.ClosetDetailResDto;
-import com.sluv.server.domain.closet.dto.ClosetItemSelectReqDto;
 import com.sluv.server.domain.closet.dto.ClosetListCountResDto;
 import com.sluv.server.domain.closet.dto.ClosetNameCheckResDto;
 import com.sluv.server.domain.closet.dto.ClosetReqDto;
@@ -12,20 +10,12 @@ import com.sluv.server.domain.closet.enums.ClosetStatus;
 import com.sluv.server.domain.closet.exception.BasicClosetDeleteException;
 import com.sluv.server.domain.closet.exception.ClosetNotFoundException;
 import com.sluv.server.domain.closet.repository.ClosetRepository;
-import com.sluv.server.domain.item.dto.ItemSimpleResDto;
-import com.sluv.server.domain.item.entity.Item;
-import com.sluv.server.domain.item.entity.ItemScrap;
-import com.sluv.server.domain.item.exception.ItemNotFoundException;
-import com.sluv.server.domain.item.repository.ItemImgRepository;
-import com.sluv.server.domain.item.repository.ItemRepository;
 import com.sluv.server.domain.item.repository.ItemScrapRepository;
 import com.sluv.server.domain.user.entity.User;
 import com.sluv.server.domain.user.exception.UserNotMatchedException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -35,13 +25,32 @@ import org.springframework.transaction.annotation.Transactional;
 public class ClosetService {
     private final ClosetRepository closetRepository;
     private final ItemScrapRepository itemScrapRepository;
-    private final ItemRepository itemRepository;
-    private final ItemImgRepository itemImgRepository;
 
+    /**
+     * 유저의 옷장 목록 조회
+     */
+    @Transactional(readOnly = true)
+    public ClosetListCountResDto getClosetList(User user) {
+        List<ClosetResDto> closetResDtos = closetRepository.getUserClosetList(user);
+        return ClosetListCountResDto.of(closetRepository.countByUserId(user.getId()), closetResDtos);
+    }
+
+    /**
+     * 옷장 생성 전 이름 중복 체크
+     */
+    @Transactional(readOnly = true)
+    public ClosetNameCheckResDto checkClosetNameDuplicated(String name, Long closetId) {
+        Boolean isDuplicated = closetRepository.checkDuplicate(name, closetId);
+
+        return ClosetNameCheckResDto.of(isDuplicated);
+    }
+
+    /**
+     * 신규 유저 생성 시 기본 옷장 생성
+     */
     @Transactional
     public void postBasicCloset(User user) {
-
-        Closet defCloset = Closet.builder()
+        Closet basicCloset = Closet.builder()
                 .user(user)
                 .name("기본 옷장")
                 .coverImgUrl(null)
@@ -50,9 +59,12 @@ public class ClosetService {
                 .closetStatus(ClosetStatus.PRIVATE)
                 .build();
 
-        closetRepository.save(defCloset);
+        closetRepository.save(basicCloset);
     }
 
+    /**
+     * 옷장 생성
+     */
     @Transactional
     public void postCloset(User user, ClosetReqDto dto) {
         log.info("옷장 생성 - 사용자: {}, 이름: {}", user.getId(), dto.getName());
@@ -61,6 +73,9 @@ public class ClosetService {
         );
     }
 
+    /**
+     * 옷장 정보 변경
+     */
     @Transactional
     public void patchCloset(User user, Long closetId, ClosetReqDto dto) {
         Closet closet = closetRepository.findById(closetId).orElseThrow(ClosetNotFoundException::new);
@@ -72,6 +87,9 @@ public class ClosetService {
         closet.changeCloset(dto);
     }
 
+    /**
+     * 옷장 삭제
+     */
     @Transactional
     public void deleteCloset(User user, Long closetId) {
         Closet closet = closetRepository.findById(closetId).orElseThrow(ClosetNotFoundException::new);
@@ -94,112 +112,4 @@ public class ClosetService {
         closetRepository.deleteById(closet.getId());
     }
 
-    @Transactional
-    public void postItemScrapToCloset(User user, Long itemId, Long closetId) {
-        Item item = itemRepository.findById(itemId).orElseThrow(ItemNotFoundException::new);
-        Closet closet = closetRepository.findById(closetId).orElseThrow(ClosetNotFoundException::new);
-
-        if (!closet.getUser().getId().equals(user.getId())) {
-            log.info("User did Not Matched. User Id: {}, Closet Owner Id : {}", user.getId(), closet.getUser().getId());
-            throw new UserNotMatchedException();
-        }
-
-        log.info("Save ItemScrap with item Id: {}, Closet Id {}", item.getId(), closet.getId());
-        ItemScrap saveItemScrap = itemScrapRepository.save(
-                ItemScrap.toEntity(item, closet)
-        );
-        log.info("Save Success with ItemScrap Id: {}", saveItemScrap.getId());
-
-        item.decreaseViewNum();
-
-    }
-
-    @Transactional
-    public void patchItems(User user, Long closetId, ClosetItemSelectReqDto dto) {
-        Closet closet = closetRepository.findById(closetId).orElseThrow(ClosetNotFoundException::new);
-
-        if (!closet.getUser().getId().equals(user.getId())) {
-            log.info("User did Not Matched. User Id: {}, Closet Owner Id : {}", user.getId(), closet.getUser().getId());
-            throw new UserNotMatchedException();
-        }
-
-        dto.getItemList().forEach(itemId -> {
-            log.info("Delete Item {}, from Closet {}", itemId, closet.getId());
-            itemScrapRepository.deleteByClosetIdAndItemId(closet.getId(), itemId);
-        });
-
-    }
-
-    @Transactional
-    public void deleteItemScrapFromCloset(User user, Long itemId) {
-        List<Closet> closetList = closetRepository.findAllByUserId(user.getId());
-
-        log.info("DELETE ItemScrap {}", itemId);
-        closetList.forEach(closet -> {
-            itemScrapRepository.deleteByClosetIdAndItemId(closet.getId(), itemId);
-        });
-
-    }
-
-    @Transactional
-    public void patchSaveCloset(User user, Long fromClosetId, Long toClosetId, ClosetItemSelectReqDto dto) {
-        Closet fromCloset = closetRepository.findById(fromClosetId).orElseThrow(ClosetNotFoundException::new);
-        Closet toCloset = closetRepository.findById(toClosetId).orElseThrow(ClosetNotFoundException::new);
-
-        // fromCloset과 현재 유저 일치 비교
-        if (!fromCloset.getUser().getId().equals(user.getId())) {
-            log.info("User did Not Matched. User Id: {}, fromCloset Owner Id : {}", user.getId(),
-                    fromCloset.getUser().getId());
-            throw new UserNotMatchedException();
-        }
-
-        // toCloset과 현재 유저 일치 비교
-        if (!toCloset.getUser().getId().equals(user.getId())) {
-            log.info("User did Not Matched. User Id: {}, toCloset Owner Id : {}", user.getId(),
-                    toCloset.getUser().getId());
-            throw new UserNotMatchedException();
-        }
-
-        // Target ItemScrap 모두 조회
-        List<ItemScrap> itemScrapList = dto.getItemList()
-                .stream()
-                .map(itemId ->
-                        itemScrapRepository.findByClosetIdAndItemId(fromClosetId, itemId)
-                ).toList();
-
-        // Target ItemScrap의 Closet을 toCloset으로 모두 변경
-        itemScrapList.forEach(itemScrap ->
-                itemScrap.changeCloset(toCloset)
-        );
-
-    }
-
-    @Transactional(readOnly = true)
-    public ClosetDetailResDto<ItemSimpleResDto> getClosetDetails(User user, Long closetId, Pageable pageable) {
-        Closet closet = closetRepository.findById(closetId).orElseThrow(ClosetNotFoundException::new);
-        if (!closet.getUser().getId().equals(user.getId())) {
-            log.info("User did Not Matched. User Id: {}, Closet Owner Id : {}", user.getId(),
-                    closet.getUser().getId());
-            throw new UserNotMatchedException();
-        }
-        Page<ItemSimpleResDto> itemPage = itemRepository.getClosetItems(closet, pageable);
-        log.info("아아아악: {}", itemPage.getContent().size());
-        return ClosetDetailResDto.of(itemPage, closet);
-    }
-
-    @Transactional(readOnly = true)
-    public ClosetListCountResDto getClosetList(User user) {
-        List<ClosetResDto> closetResDtos = closetRepository.getUserClosetList(user);
-        return ClosetListCountResDto.of(closetRepository.countByUserId(user.getId()), closetResDtos);
-    }
-
-    /**
-     * 옷장 생성 전 이름 중복 체크 기능
-     */
-    @Transactional(readOnly = true)
-    public ClosetNameCheckResDto checkClosetNameDuplicated(String name, Long closetId) {
-        Boolean isDuplicated = closetRepository.checkDuplicate(name, closetId);
-
-        return ClosetNameCheckResDto.of(isDuplicated);
-    }
 }
